requirements.txt
arrow>=0.10.0
coloredlogs>=7.3
CouchDB>=1.1
hbmqtt>=0.9.4
jsonschema>=2.6.0
Mastodon.py>=1.1.1
paho-mqtt>=1.2
pyserial>=3.2.1
pytest>=3.2.3
requests>=2.18.4
uptime>=3.0.1
testfixtures>=5.3.0
tinydb>=3.6.0
verboselogs>=1.7